Update user ID type from int to int32 in user retrieval and registration functions

+++ internal/repository/user/get_user_by_id.go
@@ -8,8 +8,8 @@ import (
 )
 
 // GetUserByID retrieves a user by their ID.
-func (i impl) GetUserByID(ctx context.Context, id int) (model.User, error) {
-	user, err := i.entClient.User.Get(ctx, int32(id))
+func (i impl) GetUserByID(ctx context.Context, id int32) (model.User, error) {
+	user, err := i.entClient.User.Get(ctx, id)
 	if err != nil {
 		return model.User{}, pkgerrors.WithStack(ErrDatabase)
 	}
